refactor(client): separate time and ui states; update types, exports, and imports (#118)
