+++ src/lib/stores/index.ts
@@ -3,6 +3,7 @@ export { streams } from './streams.svelte';
 export { createWidget } from './widget.svelte';
 export { createHomeState, homeState } from './home.svelte';
 export { createHealthState, healthState } from './health.svelte';
-export { timeState, uiState } from './ui.svelte';
+export { uiState } from './ui.svelte';
+export { timeState } from './time.svelte';
 export { musicState } from './music.svelte';
 export * from './weatherLeg';
