+++ src/lib/stores/time.types.ts
@@ -1,0 +1,1 @@
+export type QuarterHours = 1 | 2 | 3 | 4;
